+++ CHANGELOG.md
@@ -24,6 +24,7 @@ All notable changes to Rainbow-Node-SDK will be documented in this file.
 -   Add BubblesService::archiveBubble method to  close the room in one step. The other alternative is to change the status for each room users not deactivated yet. All users currently having the status 'invited' or 'accepted' will receive a message/stanza .
 -   Typescript improvement.
 -   Add CDD methods in TelephonyService : logon, logoff, withdrawal, wrapup.
+-   Add the ability to login with a token to the Rainbow Node SDK (token parameter to SDK start method.).
 
 ## [1.64.2] - 2019-11-26
 -   rebundle of the 1.64.0 version with the same content.